Add inventory Add delivery Add input Add output

+++ OamCake.Data/Dto/CakeCreationDto.cs
@@ -5,11 +5,11 @@ namespace OamCake.Data.Dto
 {
     public class CakeCreationDto
     {
-        public short? Id { get; set; }
+        public long? Id { get; set; }
         [Required(ErrorMessage = "El nombre es requerido")]
         public string Name { get; set; }
         [Required(ErrorMessage = "La foto es requerida")]
         public IFormFile? Photo { get; set; }
-        public short? CategoryId { get; set; }
+        public long? CategoryId { get; set; }
     }
 }
