+++ OamCake.Entity/CatalogDetails.cs
@@ -3,7 +3,7 @@
     public class CatalogDetails : BaseEntity
     {
         public long Id { get; set; }
-        public short CakeId { get; set; }
+        public long CakeId { get; set; }
         public long CatalogId { get; set; }
         public decimal Price { get; set; }
         public string Photo { get; set; }
