+++ OamCake.Entity/Category.cs
@@ -2,7 +2,7 @@
 {
     public class Category : BaseEntity
     {
-        public short Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
     }
 }
