+++ OamCake.Entity/CustomCake.cs
@@ -3,7 +3,7 @@
     public class CustomCake : BaseEntity
     {
         public long Id { get; set; }
-        public short? CakeId { get; set; }
+        public long? CakeId { get; set; }
         public string Photo { get; set; }
         public string Detail { get; set; }
