+++ OamCake.Entity/Ingredient.cs
@@ -5,8 +5,8 @@
         public short Id { get; set; }
         public byte Quantity { get; set; }
         public char Unid { get; set; }
-        public short? ProductId { get; set; }
-        public short CakeId { get; set; }
+        public long? ProductId { get; set; }
+        public long CakeId { get; set; }
 
         public Product? Product { get; set; }
         public Cake Cake { get; set; }
