+++ OamCake.Entity/InventoryProvider.cs
@@ -6,7 +6,7 @@
         public string Description { get; set; }
         public char Mode { get; set; }
         public decimal TotalPrice { get; set; }
-        public short ProviderId { get; set; }
+        public long ProviderId { get; set; }
 
         public Provider Provider { get; set; }
     }
