+++ OamCake.Entity/MadeCake.cs
@@ -3,7 +3,7 @@
     public class MadeCake : BaseEntity
     {
         public long Id { get; set; }
-        public short CakeId { get; set; }
+        public long CakeId { get; set; }
         public DateTime Date { get; set; }
         public char Status { get; set; }
     }
