+++ OamCake.Entity/MenuPermit.cs
@@ -2,9 +2,9 @@
 {
     public class MenuPermit : BaseEntity
     {
-        public int Id { get; set; }
-        public short MenuId { get; set; }
-        public short PermitId { get; set; }
+        public long Id { get; set; }
+        public long MenuId { get; set; }
+        public long PermitId { get; set; }
 
         public Menu Menu { get; set; }
         public Permit Permit { get; set; }
