+++ OamCake.Entity/OrderDelivery.cs
@@ -4,7 +4,7 @@
     {
         public long Id { get; set; }
         public long OrderId { get; set; }
-        public short DeliveryId { get; set; }
+        public long DeliveryId { get; set; }
         public char Status { get; set; }
         public DateTime CreatedDate { get; set; }
         public DateTime UpdatedDate { get; set; }
