+++ OamCake.Entity/ProjectionDetail.cs
@@ -3,9 +3,9 @@
     public class ProjectionDetail : BaseEntity
     {
         public long Id { get; set; }
-        public short CakeId { get; set; }
+        public long CakeId { get; set; }
         public long ProjectionId { get; set; }
-        public short Quantity { get; set; }
+        public int Quantity { get; set; }
 
         public Cake Cake { get; set; }
         public Projection Projection { get; set; }
