+++ OamCake.Entity/Provider.cs
@@ -2,7 +2,7 @@
 {
     public class Provider : BaseEntity
     {
-        public short Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Address { get; set; }
         public string Phone { get; set; }
