+++ OamCake.Entity/User.cs
@@ -2,8 +2,8 @@
 {
     public class User : BaseEntity
     {
-        public int Id { get; set; }
-        public int EmployeeId { get; set; }
+        public long Id { get; set; }
+        public long EmployeeId { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
         public bool IsActive { get; set; }
