+++ OamCake.Entity/UserRol.cs
@@ -2,9 +2,9 @@
 {
     public class UserRol : BaseEntity
     {
-        public int Id { get; set; }
-        public byte RoleId { get; set; }
-        public int UserId { get; set; }
+        public long Id { get; set; }
+        public long RoleId { get; set; }
+        public long UserId { get; set; }
         public DateTime AssignedDate { get; set; }
 
         public User User { get; set; }
